ml/requirements.camera.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
requests==2.31.0